backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
psycopg[binary]==3.2.12
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0
python-dotenv==1.0.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
python-multipart==0.0.9

# Redis & Celery
redis==5.0.0
celery==5.3.6

# Security
slowapi==0.1.9
bleach==6.1.0

# Monitoring
sentry-sdk[fastapi]==1.39.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

